backend/admin/requirements.txt
Flask==3.0.2
Flask-PyMongo==2.3.0
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.1
pymongo==4.6.1
markdown==3.5.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-docx==1.1.0
mammoth==1.7.0
beautifulsoup4==4.12.3
docx2md==0.1.1 